Simplify NODE_ENV checks so webpack can tree-shake properly
Webpack apparently recognizes references to process.env.NODE_ENV and
converts them to literals like "production", but it doesn't make the
logical leap to `typeof process === "undefined"` being reducible to
`false`. Because of this, tests weren't being stripped out of production
builds correctly.

Since AFAIK all bundlers and ESM servers have to deal with process.env
in node modules, the only time when `process` will be referenced at
runtime is when taste is running unbundled--that is, when taste's own
tests are running in the browser. We therefore patch that case by
defining a `process` global in taste's index.html file.